# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: The Canine Shelter API
  description: A simple dog listing API
  contact:
    name: Syam Marcus
    email: marcuss@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://ecology-fiction-3000.codio-box.uk/TCS/'
    description: Development server
tags:
- name: Listings
  description: Access and perform CRUD operations on dog listings
- name: Users
  description: API endpoints for user management
paths:
  /listings:
    get:
      tags:
        - Listings
      description:
        Get all dog listings
      responses:
        '200':
          description: Successful retrieval of all dog listings
          content:
            application/json:
              schema:
                $ref: ./listing.json#/definitions/listing
        '404':
          description: Unsuccessful retrieval of all dog listings
    post:
      tags:
        - Listings
      description:
        Create a new dog listing
      requestBody:
        required: true
        description: All new listing data.
        content:
          application/json:
            schema:
              $ref: ./listing.json#/definitions/listing
      responses:
        '201':
          description: Upload created listing after validation
          content:
            application/json:
              schema:
                $ref: ./listing.json#/definitions/listingCreated
        '400':
          description: Failed validation and did not upload created listing 
  /listings/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the listing
    get:
      tags:
        - Listings
      description:
        Get dog listings by listings ID
      responses:
        '200':
          description: Successful retrieval of specified dog listings
          content:
            application/json:
              schema:
                $ref: ./listing.json#/definitions/listing
        '404':
          description: Unsuccessful retrieval of specified dog listings
    put:
      tags:
        - Listings
      description:
        Update dog listing specified by ID
      requestBody:
        description: New information to update specified listing with.
        required: true
        content:
          application/json:
            schema:
              $ref: ./listing.json#/definitions/listing
      responses:
        '201':
          description: Successful update of a listing after validation
          content:
            application/json:
              schema:
                $ref: ./listing.json#/definitions/listingCreated
        '400':
          description: Failed validation and did not update listing 
            
  /register:
    get:
      tags:
        - Users
      description:
        Get list of all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/user
        '403':
          description: user is not authorised to retrieve list of users
    post:
      tags:
        - Users
      description:
        Create a new user account
      requestBody:
        description: All new user data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/user
      responses:
        '201':
          description: Successfully create a new user after validation
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/userCreated
        '400':
          description: Failed validation and did not upload created new user 